+++ tests/dsp/firmware_update.cpp
@@ -4496,3 +4496,356 @@ TEST(CancelUpdate, errorPathDecodeResponse)
         &nonFunctioningComponentIndication, &nonFunctioningComponentBitmap);
     EXPECT_EQ(rc, PLDM_ERROR_INVALID_DATA);
 }
+
+#ifdef LIBPLDM_API_TESTING
+static void compareByteVector(const uint8_t* actual, size_t actualLen,
+                              const std::vector<uint8_t>& expected)
+{
+    ASSERT_EQ(actualLen, expected.size());
+    for (size_t i = 0; i < actualLen; i++)
+    {
+        EXPECT_EQ(actual[i], expected[i]);
+    }
+}
+#endif
+
+#ifdef LIBPLDM_API_TESTING
+TEST(PackageParser, ValidPkgSingleDescriptorSingleComponent)
+{
+    std::vector<uint8_t> fwPkgHdr{
+        0xF0, 0x18, 0x87, 0x8C, 0xCB, 0x7D, 0x49, 0x43, 0x98, 0x00, 0xA0, 0x2F,
+        0x05, 0x9A, 0xCA, 0x02, 0x01, 0x8B, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
+        0x00, 0x00, 0x00, 0x19, 0x0C, 0xE5, 0x07, 0x00, 0x08, 0x00, 0x01, 0x0E,
+        0x56, 0x65, 0x72, 0x73, 0x69, 0x6F, 0x6E, 0x53, 0x74, 0x72, 0x69, 0x6E,
+        0x67, 0x31, 0x01, 0x2E, 0x00, 0x01, 0x01, 0x00, 0x00, 0x00, 0x01, 0x0E,
+        0x00, 0x00, 0x01, 0x56, 0x65, 0x72, 0x73, 0x69, 0x6F, 0x6E, 0x53, 0x74,
+        0x72, 0x69, 0x6E, 0x67, 0x32, 0x02, 0x00, 0x10, 0x00, 0x16, 0x20, 0x23,
+        0xC9, 0x3E, 0xC5, 0x41, 0x15, 0x95, 0xF4, 0x48, 0x70, 0x1D, 0x49, 0xD6,
+        0x75, 0x01, 0x00, 0x0A, 0x00, 0x64, 0x00, 0xFF, 0xFF, 0xFF, 0xFF, 0x00,
+        0x00, 0x00, 0x00, 0x8B, 0x00, 0x00, 0x00, 0x1B, 0x00, 0x00, 0x00, 0x01,
+        0x0E, 0x56, 0x65, 0x72, 0x73, 0x69, 0x6F, 0x6E, 0x53, 0x74, 0x72, 0x69,
+        0x6E, 0x67, 0x33, 0x4F, 0x96, 0xAE, 0x56};
+
+    constexpr std::string_view pkgVersion{"VersionString1"};
+
+    pldm_fwup_package_data* pkgData = nullptr;
+    int rc =
+        pldm_fwup_parse_package(fwPkgHdr.data(), fwPkgHdr.size(), &pkgData);
+    EXPECT_EQ(rc, 0);
+    ASSERT_NE(pkgData, nullptr);
+    EXPECT_EQ(pkgData->pkg_header.package_header_size, fwPkgHdr.size());
+    ASSERT_NE(pkgData->package_version_str, nullptr);
+    EXPECT_EQ(std::string_view(pkgData->package_version_str), pkgVersion);
+    EXPECT_EQ(pkgData->device_id_records_count, 1u);
+
+    auto& firstRecord = pkgData->device_id_records[0];
+    ASSERT_NE(firstRecord.comp_image_set_version, nullptr);
+    EXPECT_STREQ(firstRecord.comp_image_set_version, "VersionString2");
+    EXPECT_EQ(firstRecord.descriptors.desc_count, 1u);
+    EXPECT_EQ(pkgData->comp_image_infos_count, 1u);
+
+    auto& firstComp = pkgData->comp_image_infos[0];
+    ASSERT_NE(firstComp.comp_version, nullptr);
+    EXPECT_STREQ(firstComp.comp_version, "VersionString3");
+
+    pldm_fwup_free_package_data(pkgData);
+}
+#endif
+
+#ifdef LIBPLDM_API_TESTING
+TEST(PackageParser, ValidPkgMultipleDescriptorsMultipleComponents)
+{
+    std::vector<uint8_t> fwPkgHdr{
+        0xF0, 0x18, 0x87, 0x8C, 0xCB, 0x7D, 0x49, 0x43, 0x98, 0x00, 0xA0, 0x2F,
+        0x05, 0x9A, 0xCA, 0x02, 0x01, 0x46, 0x01, 0x00, 0x00, 0x00, 0x00, 0x00,
+        0x00, 0x00, 0x00, 0x19, 0x0C, 0xE5, 0x07, 0x00, 0x08, 0x00, 0x01, 0x0E,
+        0x56, 0x65, 0x72, 0x73, 0x69, 0x6F, 0x6E, 0x53, 0x74, 0x72, 0x69, 0x6E,
+        0x67, 0x31, 0x03, 0x45, 0x00, 0x03, 0x01, 0x00, 0x00, 0x00, 0x01, 0x0E,
+        0x00, 0x00, 0x03, 0x56, 0x65, 0x72, 0x73, 0x69, 0x6F, 0x6E, 0x53, 0x74,
+        0x72, 0x69, 0x6E, 0x67, 0x32, 0x02, 0x00, 0x10, 0x00, 0x12, 0x44, 0xD2,
+        0x64, 0x8D, 0x7D, 0x47, 0x18, 0xA0, 0x30, 0xFC, 0x8A, 0x56, 0x58, 0x7D,
+        0x5B, 0x01, 0x00, 0x04, 0x00, 0x47, 0x16, 0x00, 0x00, 0xFF, 0xFF, 0x0B,
+        0x00, 0x01, 0x07, 0x4F, 0x70, 0x65, 0x6E, 0x42, 0x4D, 0x43, 0x12, 0x34,
+        0x2E, 0x00, 0x01, 0x00, 0x00, 0x00, 0x00, 0x01, 0x0E, 0x00, 0x00, 0x07,
+        0x56, 0x65, 0x72, 0x73, 0x69, 0x6F, 0x6E, 0x53, 0x74, 0x72, 0x69, 0x6E,
+        0x67, 0x33, 0x02, 0x00, 0x10, 0x00, 0x12, 0x44, 0xD2, 0x64, 0x8D, 0x7D,
+        0x47, 0x18, 0xA0, 0x30, 0xFC, 0x8A, 0x56, 0x58, 0x7D, 0x5C, 0x2E, 0x00,
+        0x01, 0x00, 0x00, 0x00, 0x00, 0x01, 0x0E, 0x00, 0x00, 0x01, 0x56, 0x65,
+        0x72, 0x73, 0x69, 0x6F, 0x6E, 0x53, 0x74, 0x72, 0x69, 0x6E, 0x67, 0x34,
+        0x02, 0x00, 0x10, 0x00, 0x12, 0x44, 0xD2, 0x64, 0x8D, 0x7D, 0x47, 0x18,
+        0xA0, 0x30, 0xFC, 0x8A, 0x56, 0x58, 0x7D, 0x5D, 0x03, 0x00, 0x0A, 0x00,
+        0x64, 0x00, 0xFF, 0xFF, 0xFF, 0xFF, 0x00, 0x00, 0x00, 0x00, 0x46, 0x01,
+        0x00, 0x00, 0x1B, 0x00, 0x00, 0x00, 0x01, 0x0E, 0x56, 0x65, 0x72, 0x73,
+        0x69, 0x6F, 0x6E, 0x53, 0x74, 0x72, 0x69, 0x6E, 0x67, 0x35, 0x0A, 0x00,
+        0xC8, 0x00, 0xFF, 0xFF, 0xFF, 0xFF, 0x00, 0x00, 0x01, 0x00, 0x61, 0x01,
+        0x00, 0x00, 0x1B, 0x00, 0x00, 0x00, 0x01, 0x0E, 0x56, 0x65, 0x72, 0x73,
+        0x69, 0x6F, 0x6E, 0x53, 0x74, 0x72, 0x69, 0x6E, 0x67, 0x36, 0x10, 0x00,
+        0x2C, 0x01, 0xFF, 0xFF, 0xFF, 0xFF, 0x01, 0x00, 0x0C, 0x00, 0x7C, 0x01,
+        0x00, 0x00, 0x1B, 0x00, 0x00, 0x00, 0x01, 0x0E, 0x56, 0x65, 0x72, 0x73,
+        0x69, 0x6F, 0x6E, 0x53, 0x74, 0x72, 0x69, 0x6E, 0x67, 0x37, 0xF1, 0x90,
+        0x9C, 0x71};
+    pldm_fwup_package_data* pkgData = nullptr;
+    constexpr std::string_view pkgVersion{"VersionString1"};
+    int rc =
+        pldm_fwup_parse_package(fwPkgHdr.data(), fwPkgHdr.size(), &pkgData);
+    EXPECT_EQ(rc, 0);
+    ASSERT_NE(pkgData, nullptr);
+    EXPECT_EQ(pkgData->pkg_header.package_header_size, fwPkgHdr.size());
+    ASSERT_NE(pkgData->package_version_str, nullptr);
+    EXPECT_EQ(std::string_view(pkgData->package_version_str), pkgVersion);
+    ASSERT_EQ(pkgData->device_id_records_count, 3u);
+    {
+        const auto& rec = pkgData->device_id_records[0];
+        EXPECT_EQ(rec.device_update_option_flags, 1u);
+        ASSERT_EQ(rec.num_applicable_components, 2u);
+        EXPECT_EQ(rec.applicable_components[0], 0u);
+        EXPECT_EQ(rec.applicable_components[1], 1u);
+        ASSERT_NE(rec.comp_image_set_version, nullptr);
+        EXPECT_STREQ(rec.comp_image_set_version, "VersionString2");
+        EXPECT_EQ(rec.descriptors.desc_count, 3u);
+        {
+            const auto& d = rec.descriptors.desc_array[0];
+            EXPECT_EQ(d.descriptor_type, PLDM_FWUP_UUID);
+            std::vector<uint8_t> expected_uuid{
+                0x12, 0x44, 0xD2, 0x64, 0x8D, 0x7D, 0x47, 0x18,
+                0xA0, 0x30, 0xFC, 0x8A, 0x56, 0x58, 0x7D, 0x5B};
+            compareByteVector(d.descriptor_data, d.descriptor_length,
+                              expected_uuid);
+        }
+        {
+            const auto& d = rec.descriptors.desc_array[1];
+            EXPECT_EQ(d.descriptor_type, PLDM_FWUP_IANA_ENTERPRISE_ID);
+            std::vector<uint8_t> expected_iana{0x47, 0x16, 0x00, 0x00};
+            compareByteVector(d.descriptor_data, d.descriptor_length,
+                              expected_iana);
+        }
+        {
+            const auto& d = rec.descriptors.desc_array[2];
+            EXPECT_EQ(d.descriptor_type, PLDM_FWUP_VENDOR_DEFINED);
+            EXPECT_EQ(d.is_vendor_defined, true);
+            std::cout << d.vendor_title;
+            EXPECT_EQ(std::string(d.vendor_title), "OpenBMC");
+            std::vector<uint8_t> expected_iana{0x12, 0x34};
+            compareByteVector(d.descriptor_data, d.descriptor_length,
+                              expected_iana);
+        }
+    }
+    {
+        const auto& rec = pkgData->device_id_records[1];
+        EXPECT_EQ(rec.device_update_option_flags, 0u);
+        ASSERT_EQ(rec.num_applicable_components, 3u);
+        EXPECT_EQ(rec.applicable_components[0], 0u);
+        EXPECT_EQ(rec.applicable_components[1], 1u);
+        EXPECT_EQ(rec.applicable_components[2], 2u);
+        ASSERT_NE(rec.comp_image_set_version, nullptr);
+        EXPECT_STREQ(rec.comp_image_set_version, "VersionString3");
+        EXPECT_EQ(rec.descriptors.desc_count, 1u);
+        {
+            const auto& d = rec.descriptors.desc_array[0];
+            EXPECT_EQ(d.descriptor_type, PLDM_FWUP_UUID);
+            std::vector<uint8_t> expected_uuid{
+                0x12, 0x44, 0xD2, 0x64, 0x8D, 0x7D, 0x47, 0x18,
+                0xA0, 0x30, 0xFC, 0x8A, 0x56, 0x58, 0x7D, 0x5C};
+            compareByteVector(d.descriptor_data, d.descriptor_length,
+                              expected_uuid);
+        }
+    }
+    {
+        const auto& rec = pkgData->device_id_records[2];
+        EXPECT_EQ(rec.device_update_option_flags, 0u);
+        ASSERT_EQ(rec.num_applicable_components, 1u);
+        EXPECT_EQ(rec.applicable_components[0], 0u);
+        ASSERT_NE(rec.comp_image_set_version, nullptr);
+        EXPECT_STREQ(rec.comp_image_set_version, "VersionString4");
+        EXPECT_EQ(rec.descriptors.desc_count, 1u);
+        {
+            const auto& d = rec.descriptors.desc_array[0];
+            EXPECT_EQ(d.descriptor_type, PLDM_FWUP_UUID);
+            std::vector<uint8_t> expected_uuid{
+                0x12, 0x44, 0xD2, 0x64, 0x8D, 0x7D, 0x47, 0x18,
+                0xA0, 0x30, 0xFC, 0x8A, 0x56, 0x58, 0x7D, 0x5D};
+            compareByteVector(d.descriptor_data, d.descriptor_length,
+                              expected_uuid);
+        }
+    }
+
+    auto& comp_infos = pkgData->comp_image_infos;
+    auto comp_count = pkgData->comp_image_infos_count;
+    ASSERT_EQ(comp_count, 3u);
+    {
+        const auto& cinfo = comp_infos[0];
+        EXPECT_EQ(cinfo.comp_classification, 10u);
+        EXPECT_EQ(cinfo.comp_identifier, 100u);
+        EXPECT_EQ(cinfo.comp_comparison_stamp, 0xFFFFFFFFu);
+        EXPECT_EQ(cinfo.comp_options, 0u);
+        EXPECT_EQ(cinfo.requested_comp_activation_method, 0u);
+        EXPECT_EQ(cinfo.comp_location_offset, 326u);
+        EXPECT_EQ(cinfo.comp_size, 27u);
+        ASSERT_NE(cinfo.comp_version, nullptr);
+        EXPECT_STREQ(cinfo.comp_version, "VersionString5");
+    }
+    {
+        const auto& cinfo = comp_infos[1];
+        EXPECT_EQ(cinfo.comp_classification, 10u);
+        EXPECT_EQ(cinfo.comp_identifier, 200u);
+        EXPECT_EQ(cinfo.comp_comparison_stamp, 0xFFFFFFFFu);
+        EXPECT_EQ(cinfo.comp_options, 0u);
+        EXPECT_EQ(cinfo.requested_comp_activation_method, 1u);
+        EXPECT_EQ(cinfo.comp_location_offset, 353u);
+        EXPECT_EQ(cinfo.comp_size, 27u);
+        ASSERT_NE(cinfo.comp_version, nullptr);
+        EXPECT_STREQ(cinfo.comp_version, "VersionString6");
+    }
+    {
+        const auto& cinfo = comp_infos[2];
+        EXPECT_EQ(cinfo.comp_classification, 16u);
+        EXPECT_EQ(cinfo.comp_identifier, 300u);
+        EXPECT_EQ(cinfo.comp_comparison_stamp, 0xFFFFFFFFu);
+        EXPECT_EQ(cinfo.comp_options, 1u);
+        EXPECT_EQ(cinfo.requested_comp_activation_method, 12u);
+        EXPECT_EQ(cinfo.comp_location_offset, 380u);
+        EXPECT_EQ(cinfo.comp_size, 27u);
+        ASSERT_NE(cinfo.comp_version, nullptr);
+        EXPECT_STREQ(cinfo.comp_version, "VersionString7");
+    }
+    pldm_fwup_free_package_data(pkgData);
+}
+#endif
+
+#ifdef LIBPLDM_API_TESTING
+TEST(PackageParser, InvalidPkgHeaderInfoIncomplete)
+{
+    std::vector<uint8_t> fwPkgHdr{0xF0, 0x18, 0x87, 0x8C, 0xCB, 0x7D,
+                                  0x49, 0x43, 0x98, 0x00, 0xA0, 0x2F,
+                                  0x05, 0x9A, 0xCA, 0x02};
+
+    pldm_fwup_package_data* pkgData = nullptr;
+    int rc =
+        pldm_fwup_parse_package(fwPkgHdr.data(), fwPkgHdr.size(), &pkgData);
+
+    EXPECT_NE(rc, 0);
+    EXPECT_EQ(pkgData, nullptr);
+}
+#endif
+
+#ifdef LIBPLDM_API_TESTING
+TEST(PackageParser, InvalidPkgNotSupportedHeaderFormat)
+{
+    std::vector<uint8_t> fwPkgHdr{
+        0x12, 0x44, 0xD2, 0x64, 0x8D, 0x7D, 0x47, 0x18, 0xA0, 0x30,
+        0xFC, 0x8A, 0x56, 0x58, 0x7D, 0x5B, 0x02, 0x8B, 0x00, 0x00,
+        0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x19, 0x0C, 0xE5,
+        0x07, 0x00, 0x08, 0x00, 0x01, 0x0E, 0x56, 0x65, 0x72, 0x73,
+        0x69, 0x6F, 0x6E, 0x53, 0x74, 0x72, 0x69, 0x6E, 0x67, 0x31};
+
+    pldm_fwup_package_data* pkgData = nullptr;
+    int rc =
+        pldm_fwup_parse_package(fwPkgHdr.data(), fwPkgHdr.size(), &pkgData);
+    EXPECT_NE(rc, 0);
+    EXPECT_EQ(pkgData, nullptr);
+}
+#endif
+
+#ifdef LIBPLDM_API_TESTING
+class PackageParserMultipleDescSameType : public ::testing::Test
+{
+  protected:
+    PackageParserMultipleDescSameType()
+    {
+    }
+
+    std::vector<uint8_t> fwPkgHdr{
+        0xF0, 0x18, 0x87, 0x8C, 0xCB, 0x7D, 0x49, 0x43, 0x98, 0x00, 0xA0, 0x2F,
+        0x05, 0x9A, 0xCA, 0x02, 0x01, 0xB6, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
+        0x00, 0x00, 0x00, 0x19, 0x0C, 0xE5, 0x07, 0x00, 0x08, 0x00, 0x01, 0x0E,
+        0x56, 0x65, 0x72, 0x73, 0x69, 0x6F, 0x6E, 0x53, 0x74, 0x72, 0x69, 0x6E,
+        0x67, 0x31, 0x01, 0x59, 0x00, 0x05, 0x01, 0x00, 0x00, 0x00, 0x01, 0x0E,
+        0x00, 0x00, 0x01, 0x56, 0x65, 0x72, 0x73, 0x69, 0x6F, 0x6E, 0x53, 0x74,
+        0x72, 0x69, 0x6E, 0x67, 0x32, 0x02, 0x00, 0x10, 0x00, 0x12, 0x44, 0xD2,
+        0x64, 0x8D, 0x7D, 0x47, 0x18, 0xA0, 0x30, 0xFC, 0x8A, 0x56, 0x58, 0x7D,
+        0x5B, 0x01, 0x00, 0x04, 0x00, 0x47, 0x16, 0x00, 0x00, 0x01, 0x00, 0x04,
+        0x00, 0x47, 0x16, 0x10, 0x00, 0xFF, 0xFF, 0x0B, 0x00, 0x01, 0x07, 0x4F,
+        0x70, 0x65, 0x6E, 0x42, 0x4D, 0x43, 0x12, 0x34, 0xFF, 0xFF, 0x08, 0x00,
+        0x01, 0x03, 0x53, 0x4B, 0x55, 0x12, 0x34, 0x56, 0x01, 0x00, 0x0A, 0x00,
+        0x64, 0x00, 0xFE, 0xFF, 0xFF, 0xFF, 0x02, 0x00, 0x00, 0x00, 0xB6, 0x00,
+        0x00, 0x00, 0x02, 0x00, 0x00, 0x00, 0x01, 0x0E, 0x56, 0x65, 0x72, 0x73,
+        0x69, 0x6F, 0x6E, 0x53, 0x74, 0x72, 0x69, 0x6E, 0x67, 0x35, 0x11, 0x16,
+        0x1E, 0x1B};
+
+    std::string_view pkgVersion{"VersionString1"};
+
+    pldm_fwup_package_data* pkgData = nullptr;
+
+    void SetUp() override
+    {
+        int rc =
+            pldm_fwup_parse_package(fwPkgHdr.data(), fwPkgHdr.size(), &pkgData);
+        ASSERT_EQ(rc, 0);
+        ASSERT_NE(pkgData, nullptr);
+    }
+
+    void TearDown() override
+    {
+        pldm_fwup_free_package_data(pkgData);
+        pkgData = nullptr;
+    }
+};
+#endif
+
+#ifdef LIBPLDM_API_TESTING
+TEST_F(PackageParserMultipleDescSameType, DescriptorsMatch)
+{
+    EXPECT_EQ(pkgData->pkg_header.package_header_size, fwPkgHdr.size());
+    ASSERT_NE(pkgData->package_version_str, nullptr);
+    EXPECT_EQ(std::string_view(pkgData->package_version_str), pkgVersion);
+    ASSERT_EQ(pkgData->device_id_records_count, 1u);
+    const auto& rec = pkgData->device_id_records[0];
+    EXPECT_EQ(rec.device_update_option_flags, 1u);
+    ASSERT_NE(rec.comp_image_set_version, nullptr);
+    EXPECT_STREQ(rec.comp_image_set_version, "VersionString2");
+    ASSERT_EQ(rec.descriptors.desc_count, 5u);
+    {
+        const auto& d = rec.descriptors.desc_array[0];
+        EXPECT_EQ(d.descriptor_type, PLDM_FWUP_UUID);
+        std::vector<uint8_t> expectedUuid{0x12, 0x44, 0xD2, 0x64, 0x8D, 0x7D,
+                                          0x47, 0x18, 0xA0, 0x30, 0xFC, 0x8A,
+                                          0x56, 0x58, 0x7D, 0x5B};
+        compareByteVector(d.descriptor_data, d.descriptor_length, expectedUuid);
+    }
+    {
+        const auto& d = rec.descriptors.desc_array[1];
+        EXPECT_EQ(d.descriptor_type, PLDM_FWUP_IANA_ENTERPRISE_ID);
+        std::vector<uint8_t> expectedIana{0x47, 0x16, 0x00, 0x00};
+        compareByteVector(d.descriptor_data, d.descriptor_length, expectedIana);
+    }
+    {
+        const auto& d = rec.descriptors.desc_array[2];
+        EXPECT_EQ(d.descriptor_type, PLDM_FWUP_IANA_ENTERPRISE_ID);
+        std::vector<uint8_t> expectedIana2{0x47, 0x16, 0x10, 0x00};
+        compareByteVector(d.descriptor_data, d.descriptor_length,
+                          expectedIana2);
+    }
+    {
+        const auto& d = rec.descriptors.desc_array[3];
+        EXPECT_EQ(d.descriptor_type, PLDM_FWUP_VENDOR_DEFINED);
+
+        ASSERT_NE(d.vendor_title, nullptr);
+        EXPECT_STREQ(d.vendor_title, "OpenBMC");
+
+        std::vector<uint8_t> expectedVendor{0x12, 0x34};
+        compareByteVector(d.descriptor_data, d.descriptor_length,
+                          expectedVendor);
+    }
+    {
+        const auto& d = rec.descriptors.desc_array[4];
+        EXPECT_EQ(d.descriptor_type, PLDM_FWUP_VENDOR_DEFINED);
+
+        ASSERT_NE(d.vendor_title, nullptr);
+        EXPECT_STREQ(d.vendor_title, "SKU");
+
+        std::vector<uint8_t> expectedVendor{0x12, 0x34, 0x56};
+        compareByteVector(d.descriptor_data, d.descriptor_length,
+                          expectedVendor);
+    }
+}
+#endif